requirements.txt
argparse==1.4.0           # via caniusepython3
astroid==1.5.2            # via edx-lint, pylint, pylint-celery, pylint-plugin-utils
backports.functools-lru-cache==1.4  # via astroid, caniusepython3, pylint
caniusepython3==6.0.0
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
click-log==0.1.8          # via edx-lint
click==6.7                # via click-log, edx-lint, pip-tools
configparser==3.5.0       # via pylint
diff-cover==1.0.2
distlib==0.2.6            # via caniusepython3
django==2.0
djangorestframework==3.7.7
edx-i18n-tools==0.4.3
edx-lint==0.5.4
enum34==1.1.6             # via astroid
first==2.0.1              # via pip-tools
#futures==3.3.0            # via caniusepython3
idna==2.6                 # via requests
inflect==0.2.5            # via jinja2-pluralize
isort==4.2.15
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10              # via diff-cover, jinja2-pluralize
jsonfield==2.0.2
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via pylint
packaging==16.8           # via caniusepython3
path.py==10.5             # via edx-i18n-tools
pip-tools==1.11.0
pkginfo==1.4.1            # via twine
pluggy==0.6.0             # via tox
polib==1.1.0              # via edx-i18n-tools
py==1.5.2                 # via tox
pycodestyle==2.3.1
pydocstyle==2.1.1
pygments==2.2.0           # via diff-cover
pylint-celery==0.3        # via edx-lint
pylint-django==0.7.2      # via edx-lint
pylint-plugin-utils==0.2.6  # via pylint-celery, pylint-django
pylint==1.7.1             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.2.0          # via packaging
pytz==2017.3              # via django
pyyaml==3.12              # via edx-i18n-tools
requests-toolbelt==0.8.0  # via twine
requests==2.18.4          # via caniusepython3, requests-toolbelt, twine
singledispatch==3.4.0.3   # via astroid, pylint
six==1.11.0               # via astroid, diff-cover, edx-i18n-tools, edx-lint, packaging, pip-tools, pydocstyle, pylint, singledispatch, tox
snowballstemmer==1.2.1    # via pydocstyle
tox-battery==0.2
tox==2.9.1
tqdm==4.19.5              # via twine
twine==1.9.1
urllib3==1.22             # via requests
virtualenv==15.1.0        # via tox
wheel==0.30.0
wrapt==1.10.11            # via astroid